+++ packages/credential-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/credential-sdk
 
+## 0.34.0
+
+### Minor Changes
+
+- Expose `accumulatorHistory`
+
 ## 0.33.0
 
 ### Minor Changes